backend/requirements.txt
Flask
Flask-SQLAlchemy
python-dateutil
pytest